pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "eigenfraud"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "pandas>=2.2",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["eigenfraud*"]
